Split core/data_var.py off type.py
