+++ superduperdb/models/torch/wrapper.py
@@ -4,8 +4,8 @@ from typing import Optional, Callable
 import torch
 from torch.utils import data
 
+from superduperdb.core.data_var import DataVar
 from superduperdb.core.documents import Document
-from superduperdb.core.type import DataVar
 from superduperdb.misc import progress
 from superduperdb.core.model import Model
 from superduperdb.models.torch.utils import device_of, to_device, eval
